+++ dynamics/src/include/VPCGDynamicsKernel.hpp
@@ -109,8 +109,8 @@ public:
 protected:
     StressUpdateStep<DGadvection, DGstressComp>& stressStep;
     const VPParameters& params;
-    const double alpha = 1500.;
-    const double beta = 1500.;
+    const FloatType alpha = 1500.;
+    const FloatType beta = 1500.;
 
     // Step-initial ice velocity
     CGVector<CGdegree> u0;
@@ -120,8 +120,8 @@ protected:
     {
         // Update the velocity
 
-        const double FOcean = params.COcean * params.rhoOcean;
-        const double FAtm = params.CAtm * params.rhoAtm;
+        const FloatType FOcean = params.COcean * params.rhoOcean;
+        const FloatType FAtm = params.CAtm * params.rhoAtm;
 
         //      update by a loop.. implicit parts and h-dependent
 #pragma omp parallel for
@@ -130,8 +130,8 @@ protected:
                 continue;
             auto uOcnRel = u(i) - uOcean(i);
             auto vOcnRel = v(i) - vOcean(i);
-            double absatm = sqrt(SQR(uAtmos(i)) + SQR(vAtmos(i)));
-            double absocn = sqrt(
+            FloatType absatm = sqrt(SQR(uAtmos(i)) + SQR(vAtmos(i)));
+            FloatType absocn = sqrt(
                 SQR(uOcnRel) + SQR(vOcnRel)); // note that the sign of uOcnRel is irrelevant here
             auto uPrev = u(i);
